+++ src/test/java/com/andersen/pre_intensive/LinkedListImplementationTest.java
@@ -1,0 +1,36 @@
+package com.andersen.pre_intensive;
+
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+public class LinkedListImplementationTest {
+
+    @Test
+    public void add() {
+    }
+
+    @Test
+    public void testAdd() {
+    }
+
+    @Test
+    public void sort() {
+    }
+
+    @Test
+    public void concat() {
+    }
+
+    @Test
+    public void delete() {
+    }
+
+    @Test
+    public void get() {
+    }
+
+    @Test
+    public void size() {
+    }
+}
